requirements.txt
Flask
easyocr
numpy
Pillow
Werkzeug
gunicorn